backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.10.3
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
bcrypt>=5.0.0
python-jose[cryptography]
python-multipart==0.0.9
email-validator==2.1.0
python-dotenv==1.0.1
